+++ clang-tools-extra/clang-tidy/portability/PortabilityTidyModule.cpp
@@ -8,6 +8,7 @@
 
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
+#include "AvoidPlatformSpecificFundamentalTypesCheck.h"
 #include "AvoidPragmaOnceCheck.h"
 #include "NoAssemblerCheck.h"
 #include "RestrictSystemIncludesCheck.h"
@@ -22,6 +23,8 @@ namespace {
 class PortabilityModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+    CheckFactories.registerCheck<AvoidPlatformSpecificFundamentalTypesCheck>(
+        "portability-avoid-platform-specific-fundamental-types");
     CheckFactories.registerCheck<AvoidPragmaOnceCheck>(
         "portability-avoid-pragma-once");
     CheckFactories.registerCheck<NoAssemblerCheck>("portability-no-assembler");